refactor(setbuilder): extract pure helpers out of BuilderWorkspace (#532)
Move three pure helpers out of the BuilderWorkspace React component into
focused modules (behavior-preserving):

- buildReorderedIds → reorderMath.ts (alongside buildMovedIds; kept as a
  distinct parallel implementation of the locked-slot-anchor invariant —
  drag vs one-step — NOT merged)
- insertPoolTrackIntoDocument + lockSlotsInDocument → new documentMath.ts

The test now imports buildReorderedIds from reorderMath instead of the
component.

Closes #532
